Move tests into test dir
